Files with flight data in an array
I've changed the js file so the flight data is stored in an array of objects.  I'm matching the flight status from the array with the flight number entered by the user using a for loop.  in the javascript file.  I could make this code better by changing the flight numbers in the object back to integers and writing a script that converts the user's entry into a number.  As is, I've made the flight numbers in the array strings so they match the user entry.